\name{snptree-package}
\alias{snptree-package}
\alias{snptree}
\docType{package}
\title{
Split haplotypes into binary trees using lexical sorting.
}
\description{
Split haplotypes into binary trees using lexical sorting based on the order of sample.
}
\details{
\tabular{ll}{
Package: \tab snptree\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-04-25\cr
License: \tab GPL-3\cr
}
This package splits haplotypes into trees via lexical sorting.
}
\author{
  Ian Wilson <email: ian.wilson@ncl.ac.uk>
  Maintainer: Who to complain to <ian.wilson@newcastle.ac.uk>
}
\references{Finding genes that influence quantitative traits with tree-based clustering.
I. J. Wilson, R. A. J. Howey, D. T. Houniet and M. Santibanez-Koref.  BMC
Proceedings. vol. 5(supp 9):S98 2011.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
